requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-clear-cache==0.3
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
